move geometry metadata constraint to common bundle
