+++ io/plugins/eu.esdihumboldt.hale.io.jdbc.mssql/src/eu/esdihumboldt/hale/io/jdbc/mssql/MsSqlGeometries.java
@@ -41,9 +41,9 @@ import eu.esdihumboldt.hale.common.instance.geometry.impl.WKTDefinition;
 import eu.esdihumboldt.hale.common.schema.geometry.CRSDefinition;
 import eu.esdihumboldt.hale.common.schema.geometry.GeometryProperty;
 import eu.esdihumboldt.hale.common.schema.model.TypeDefinition;
+import eu.esdihumboldt.hale.common.schema.model.constraint.type.GeometryMetadata;
 import eu.esdihumboldt.hale.common.schema.model.impl.DefaultTypeDefinition;
 import eu.esdihumboldt.hale.io.jdbc.GeometryAdvisor;
-import eu.esdihumboldt.hale.io.jdbc.constraints.GeometryMetadata;
 import eu.esdihumboldt.hale.io.jdbc.mssql.util.SRSUtil;
 import schemacrawler.schema.BaseColumn;
 import schemacrawler.schema.ColumnDataType;
